+++ src/commands/generate/dataset.ts
@@ -3,12 +3,12 @@ import type { Command } from 'commander';
 import * as fs from 'fs';
 import yaml from 'js-yaml';
 import { disableCache } from '../../cache';
-import { resolveConfigs } from '../../config';
 import logger from '../../logger';
 import telemetry from '../../telemetry';
 import { synthesizeFromTestSuite } from '../../testCases';
 import type { TestSuite, UnifiedConfig } from '../../types';
 import { printBorder, setupEnv } from '../../util';
+import { resolveConfigs } from '../../util/config/load';
 
 interface DatasetGenerateOptions {
   cache: boolean;
